requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.0
trafilatura>=1.12.0
python-dotenv>=1.0.1
pydantic>=2.7.0
openai>=1.40.0
tqdm>=4.66.0

feedparser>=6.0.11
